I got the circular motion working
The next step for me would be to create an if statement that gets the
red/pink marker to move in the opposite direction when a key is
pressed.  I know the keypressed function, the challenge for me would be
getting the marker to move in the opposite direction once a key is
pressed.